Refactor user and admin controllers by removing unused profile update and change password functions; reorganize imports in user.controller.js
