+++ controller/admin.controller.js
@@ -421,66 +421,3 @@ export const deleteSellerRequest = catchAsync(async (req, res) => {
     data: "",
   });
 });
-
-// Setting (Update Profile)
-export const updateProfile = catchAsync(async (req, res) => {
-  const { fullName, username, phoneNumber, dateOfBirth, gender, address } =
-    req.body;
-  const user = await User.findById(req.user._id);
-  if (!user) {
-    throw new AppError(httpStatus.NOT_FOUND, "User not found");
-  }
-
-  if (fullName) user.fullName = fullName;
-  if (username) user.username = username;
-  if (phoneNumber) user.phoneNumber = phoneNumber;
-  if (dateOfBirth) user.dateOfBirth = dateOfBirth;
-  if (gender) user.gender = gender;
-  if (address) user.address = address;
-
-  await user.save();
-
-  sendResponse(res, {
-    statusCode: httpStatus.OK,
-    success: true,
-    message: "Profile updated successfully",
-    data: user,
-  });
-});
-
-// Change Password
-export const changePassword = catchAsync(async (req, res) => {
-  const { oldPassword, newPassword } = req.body;
-  if (!oldPassword || !newPassword) {
-    throw new AppError(
-      httpStatus.BAD_REQUEST,
-      "Old password and new password are required"
-    );
-  }
-  if (oldPassword === newPassword) {
-    throw new AppError(
-      httpStatus.BAD_REQUEST,
-      "Old password and new password cannot be same"
-    );
-  }
-  const user = await User.findById(req.user._id).select("+password");
-
-  if (!user) {
-    throw new AppError(httpStatus.NOT_FOUND, "User not found");
-  }
-
-  const isMatch = await user.isPasswordMatched(oldPassword, user.password);
-  if (!isMatch) {
-    throw new AppError(httpStatus.UNAUTHORIZED, "Old password is incorrect");
-  }
-
-  user.password = newPassword;
-  await user.save();
-
-  sendResponse(res, {
-    statusCode: httpStatus.OK,
-    success: true,
-    message: "Password changed successfully",
-    data: "",
-  });
-});
