+++ controller/user.controller.js
@@ -1,9 +1,9 @@
-import catchAsync from "../utils/catchAsync.js";
-import sendResponse from "../utils/sendResponse.js";
+import httpStatus from "http-status";
+import { User } from "../model/user.model.js";
 import { uploadOnCloudinary } from "../utils/commonMethod.js";
 import AppError from "../errors/AppError.js";
-import httpStatus from "http-status";
-import { User } from "./../model/user.model.js";
+import sendResponse from "../utils/sendResponse.js";
+import catchAsync from "../utils/catchAsync.js";
 
 // Get user profile
 export const getProfile = catchAsync(async (req, res) => {
@@ -30,6 +30,7 @@ export const updateProfile = catchAsync(async (req, res) => {
     throw new AppError(httpStatus.NOT_FOUND, "User not found");
   }
 
+  // Update fields that are included in the request body
   if (name) user.name = name;
   if (username) user.username = username;
   if (phone) user.phone = phone;
@@ -38,6 +39,7 @@ export const updateProfile = catchAsync(async (req, res) => {
   if (state) user.address.state = state;
   if (zipCode) user.address.zipCode = zipCode;
 
+  // If avatar file is uploaded, process it
   if (req.file) {
     const result = await uploadOnCloudinary(req.file.buffer, {
       folder: "avatars",
@@ -54,6 +56,39 @@ export const updateProfile = catchAsync(async (req, res) => {
     statusCode: httpStatus.OK,
     success: true,
     message: "Profile updated successfully",
+    data: user,
+  });
+});
+
+// Change user password
+export const changePassword = catchAsync(async (req, res) => {
+  const { currentPassword, newPassword, confirmPassword } = req.body;
+  const user = await User.findById(req.user._id).select("+password");
+  if (!user) {
+    throw new AppError(httpStatus.NOT_FOUND, "User not found");
+  }
+
+  if (newPassword !== confirmPassword) {
+    throw new AppError(
+      httpStatus.BAD_REQUEST,
+      "New password and confirm password do not match"
+    );
+  }
+
+  if (!(await User.isPasswordMatched(currentPassword, user.password))) {
+    throw new AppError(
+      httpStatus.UNAUTHORIZED,
+      "Current password is incorrect"
+    );
+  }
+
+  user.password = newPassword;
+  await user.save();
+
+  sendResponse(res, {
+    statusCode: httpStatus.OK,
+    success: true,
+    message: "Password changed successfully",
     data: user,
   });
 });
